% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backoffngram.R
\name{ngram_lprob}
\alias{ngram_lprob}
\title{Calculate log probability of the sequence}
\usage{
ngram_lprob(ngram, x)
}
\arguments{
\item{ngram}{An n-gram model}

\item{x}{q sequence of characters (words). It should be longer than the length of the n-gram.}
}
\value{
log probability of x
}
\description{
Calculate log probability of the sequence
}
