% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backoffngram.R
\name{ngram_train}
\alias{ngram_train}
\title{Train back-off n-gram language model}
\usage{
ngram_train(x, n, unk_thres = 1, start_sym = NA, end_sym = NA)
}
\arguments{
\item{x}{a vector of characters (words)}

\item{n}{length of n-gram}

\item{unk_thres}{threshold of word frequency.
If the frequency of the word is no more than this value, that word is replaced with UNK symbol.}

\item{start_sym}{if not NA, this symbol is appended at the top of x}

\item{end_sym}{if not NA, this symbol is appended at the end of x}
}
\value{
n-gram model
}
\description{
Train back-off n-gram language model
}
